requirements-dev.txt
-r requirements.txt
pytest>=3.6
pytest-django
mixer
